requirements.txt
pexif==0.15
Pillow==3.3.0